restructure keeper
